Supporting DELETE for nanopubs with any URI.

+++ whyis/blueprint/entity/delete_entity.py
@@ -1,6 +1,7 @@
 from flask import current_app
 from flask_login import login_required
 from whyis.decorator import conditional_login_required
+from rdflib import *
 
 from .entity_blueprint import entity_blueprint
 
@@ -10,6 +11,10 @@ from .entity_blueprint import entity_blueprint
 def delete_entity(name=None, format=None, view=None):
     current_app.db.store.nsBindings = {}
     entity, content_type = current_app.get_entity_uri(name, format)
-    
-    current_app.delete_file(entity)
+    if not current_app._can_edit(entity):
+        return '<h1>Not Authorized</h1>', 401
+    if (entity, RDF.type, current_app.NS.np.Nanopublication) in current_app.db:
+        current_app.nanopub_manager.retire(entity)
+    else:
+        current_app.delete_file(entity)
     return '', 204
